requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
configparser==3.5.0
-e git+git@github.com:RobbieSoutham/DBCF-StocksX.git@faeb41e1d67886384699cfc3fcbc5126aa90f854#egg=DBCFStocksX
enum34==1.1.6
Flask==1.0.2
Flask-Script==2.0.5
Flask-WTF==0.14.2
futures==3.2.0
gunicorn==18.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==0.23
mccabe==0.6.1
mysqlclient==1.3.14
pycparser==2.19
pylint==1.9.3
pytoml==0.1.20
PyYAML==3.13
simplejson==3.5.2
singledispatch==3.4.0.3
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xmltodict==0.11.0
yasha==4.3
